clean up types to own file, simplify hooks for loading state management only, simplified actions, resolve crypto by no edge runtimes
