+++ apps/dashboard/app/(app)/apis/[apiId]/page.tsx
@@ -19,7 +19,6 @@ import { navigation } from "./constants";
 import { type Interval, IntervalSelect } from "./select";
 
 export const dynamic = "force-dynamic";
-export const runtime = "edge";
 
 export default async function ApiPage(props: {
   params: { apiId: string };
