+++ apps/dashboard/app/(app)/authorization/layout.tsx
@@ -5,7 +5,6 @@ import { db } from "@/lib/db";
 import { redirect } from "next/navigation";
 
 export const dynamic = "force-dynamic";
-export const runtime = "edge";
 
 export default async function AuthorizationLayout({
   children,
