+++ apps/dashboard/app/(app)/ratelimits/[namespaceId]/page.tsx
@@ -3,7 +3,6 @@ import { NamespaceNavbar } from "./namespace-navbar";
 import { getWorkspaceDetails } from "./namespace.actions";
 
 export const dynamic = "force-dynamic";
-export const runtime = "edge";
 
 export default async function RatelimitNamespacePage(props: {
   params: { namespaceId: string };
