+++ apps/dashboard/app/(app)/settings/root-keys/[keyId]/page.tsx
@@ -19,7 +19,6 @@ import { Workspace } from "./permissions/workspace";
 import { UpdateRootKeyName } from "./update-root-key-name";
 
 export const dynamic = "force-dynamic";
-export const runtime = "edge";
 
 export default async function RootKeyPage(props: {
   params: { keyId: string };
