+++ apps/dashboard/app/auth/sso-callback/[[...sso-callback]]/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { auth } from '@/lib/auth/index'
+import { auth } from '@/lib/auth/server';
 import { cookies } from 'next/headers';
 
 export async function GET(request: NextRequest) {
