+++ apps/dashboard/app/layout.tsx
@@ -12,7 +12,6 @@ import type React from "react";
 import { Suspense } from "react";
 import { ReactQueryProvider } from "./react-query-provider";
 import { ThemeProvider } from "./theme-provider";
-import { AuthProvider } from '@/lib/auth/auth-provider'
 
 export const metadata = {
   metadataBase: new URL("https://unkey.dev"),
@@ -55,7 +54,6 @@ export default function RootLayout({
       <Suspense>
         <PostHogPageview />
       </Suspense>
-      <AuthProvider>
       <PHProvider>
         <body className="min-h-full antialiased">
           <Toaster />
@@ -69,7 +67,6 @@ export default function RootLayout({
             </ReactQueryProvider>
         </body>
       </PHProvider>
-      </AuthProvider>
     </html>
   );
 }
