+++ apps/dashboard/lib/auth.ts
@@ -1,4 +1,4 @@
-import { auth } from "@/lib/auth/index";
+import { auth } from "@/lib/auth/server";
 import { redirect } from "next/navigation";
 
 /**
