+++ apps/dashboard/lib/trpc/routers/workspace/create.ts
@@ -1,6 +1,6 @@
 import { insertAuditLogs } from "@/lib/audit";
 import { type Workspace, db, schema } from "@/lib/db";
-import { auth as authProvider } from "@/lib/auth/index"
+import { auth as authProvider } from "@/lib/auth/server";
 import { TRPCError } from "@trpc/server";
 import { defaultProSubscriptions } from "@unkey/billing";
 import { newId } from "@unkey/id";
