+++ apps/dashboard/pages/api/trpc/[trpc].ts
@@ -3,7 +3,6 @@ import { router } from "@/lib/trpc/routers";
 import { fetchRequestHandler } from "@trpc/server/adapters/fetch";
 import type { NextRequest } from "next/server";
 
-export const runtime = "edge";
 // export API handler
 export default async function handler(req: NextRequest) {
   return fetchRequestHandler({
